moved non-routing functions to error_functions.py
